requirements.txt
asgiref==3.5.2
Django==4.1.1
django-ninja==0.19.1
dnspython==2.2.1
email-validator==1.2.1
idna==3.3
install==1.3.5
jose==1.0.0
Pillow==9.2.0
pydantic==1.10.2
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
